requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-bootstrap3==6.2.2
django-ckeditor==5.0.2
django-taggit==0.17.6
django-toolbelt==0.0.1
gunicorn==19.4.3
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
